Use new edition templates
